# submission/api/openapi.yaml
openapi: 3.1.0
info:
  title: LabelHub Core API
  version: "1.0.0"
  summary: HTTP edge of the LabelHub data-annotation platform.
  description: >
    Portable OpenAPI description of LabelHub's **HTTP route handlers**
    (`src/app/api/**`). These thin external edges authenticate, then delegate to
    the library layer. They cover the **machine-to-machine** surfaces:
    workspace-API-key ingest/read endpoints, admin (Supabase session) export
    endpoints, public operational endpoints, and the LLM gateway proxy.


    LabelHub deliberately keeps **internal mutations off HTTP** — the browser UI
    creates tasks, submits/reviews annotations, and runs payouts through
    **Next.js Server Actions** (`'use server'` functions), NOT these routes. That
    RPC surface is not expressible in OpenAPI; it is documented in the companion
    `API.md` ("Server-Action RPC surface"). This file documents only what is
    reachable over HTTP.


    ## Authentication models

    Three distinct models are in play (see `securitySchemes`):

    1. **Workspace API key** — machine-to-machine. A `lh_ws_*` (or rate-limited
       public `lh_demo_*`) bearer token. Only the SHA-256 hash is stored; the
       workspace is *inferred from the key*, so these endpoints take no
       `workspaceId`. Accepted on `Authorization: Bearer <key>`, bare
       `Authorization: <key>`, `x-api-key: <key>`, or `x-labelhub-api-key: <key>`.
    2. **Supabase user session** — cookie session from `@supabase/ssr`. Used by
...

...
servers:
...
paths:
...
  /api/workspaces/{id}/tasks/{taskId}/export:
    get:
      tags: [Export]
      summary: Export one task's submitted annotations
      description: >
        Task-scoped annotation export (admin session auth). One row per
        submitted annotation, including AI-review fields, human-review events,
        and (trajectory mode) nested step annotations. Capped/paginated via
        `limit` (max 50000) + `offset`.
      operationId: exportTaskAnnotations
      security:
        - supabaseSession: []
      parameters:
        - name: id
          in: path
          required: true
          description: Workspace id.
          schema: { type: string, format: uuid }
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: format
          in: query
          schema: { type: string, enum: [json, jsonl, csv, excel], default: json }
        - name: mapping
          in: query
          description: JSON array of field-mapping entries (overrides the default table mapping).
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50000, maximum: 50000 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Annotation export in the requested format.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
            application/jsonl:
              schema: { type: string }
            text/csv:
              schema: { type: string }
            application/octet-stream:
              schema: { type: string, format: binary }
        "400":
          description: Unknown format (`VALIDATION_ERROR`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FlatError" }
        "401":
          description: Not signed in.
        "403":
          description: Not a workspace admin.
        "404":
          description: Task not found / not in this workspace (`NOT_FOUND`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FlatError" }
        "500":
          $ref: "#/components/responses/FlatInternalError"

  # ───────────────────────── Eval runs (session auth) ────────────────────────
...
components:
  securitySchemes:
    workspaceApiKey:
      type: http
      scheme: bearer
      description: >
        Workspace API key (`lh_ws_*`, or rate-limited public `lh_demo_*`).
        Accepted on `Authorization: Bearer <key>`, bare `Authorization: <key>`,
        `x-api-key: <key>`, or `x-labelhub-api-key: <key>`. The workspace is
        inferred from the key.
    supabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: >
        Supabase cookie session (`@supabase/ssr`). The route resolves the target
        workspace, then enforces a workspace-scoped guard
        (`requireWorkspaceAdmin` / `requireWorkspaceMember`). Cookie name varies
        by Supabase project ref.
    adminToken:
      type: http
      scheme: bearer
      description: >
        Out-of-band ops secret from the `ADMIN_DIAG_TOKEN` env. Send via
        `Authorization: Bearer <token>` or `x-admin-token`; `?token=` query is a
        soft-deprecated fallback.
  schemas:
    FlatError:
      type: object
      description: Error envelope used by ingest / export / eval-run routes.
      properties:
        error: { type: string }
        code: { type: string }
      required: [error, code]
...
  responses:
    FlatInternalError:
      description: "Internal error (generic; no stack/DB text). `code: INTERNAL`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FlatError" }
...